fix(SW-27165): Change consideration of graduated prices in order module
While adding new positions to an existing order in the backend order module, graduated prices are still considered.
If an existing position is changed, the graduated price will only be checked and considered if the quantity was changed.
With this change it is possible again to also set a custom price for a position that contains a product with graduated prices.